Add workflow history & edit feature to training events
Implemented full workflow history timeline on event detail page with backend support and new CSS. Added Edit functionality: controller action, validation, permission checks, and update logic. Updated DTOs/models for histories and workflow data. Created interactive Edit view with participant management and trainer selection. Refactored detail page to show timeline. Cleaned up obsolete JS code.

+++ Models/History.cs
@@ -2,7 +2,7 @@
 {
     public class History
     {
-            public int DocId { get; set; }
+            public long DocId { get; set; }
             public int Year { get; set; }
             public int Month { get; set; }
             public string? DocType { get; set; }
